+++ changelog.md
@@ -3,3 +3,8 @@
 ### 0.1.0 - 2017-09-20
 
 * Initial Version
+
+### 0.2.0 - 2017-09-21
+
+* Make script repository global
+* Add clone command